Demo 1: Green
We fix the functional code to make the test pass. The shortest path to make it happen here is also the final path: functional change is so simple that we don't need to fake it just to prove that the test is correct.

We also have the first test to safeguard against any bugs we may introduce by changing functional code.